requirements.txt
telethon==1.34.0
Pillow==10.2.0
aiofiles==23.2.1
click==8.1.7